requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
category==0.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.6
django-currencies==0.10.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-mptt==0.14.0
django-rosetta==0.9.9
idna==3.4
jsonfield==2.1.1
numpy==1.24.2
Pillow==9.4.0
polib==1.2.0
pony==0.7.16
python-gettext==5.0
requests==2.28.2
rust_category==0.2.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
store==2020.2.10
stripe==5.3.0
typing_extensions==4.6.0
tzdata==2022.7
urllib3==1.26.15
